requirements.txt
asgiref==3.7.2
Django==4.2.8
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
Pillow==10.1.0
PyJWT==2.8.0
pytz==2023.3.post1
sqlparse==0.4.4
typing-extensions==4.9.0
psycopg2-binary==2.8.6
drf-yasg==1.20.0